+++ backend/src/main/java/com/planify/backend/repository/UserRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<@NonNull User, @NonNull Long> {
+public interface UserRepository extends JpaRepository<@NonNull User, @NonNull Integer> {
     Optional<User> findByUsername(String username);
 }
